requirements.txt
google-generativeai==0.8.3
streamlit==1.41.1
Pillow==11.0.0
pillow-heif==0.21.0